curl options

With the opts parameter you can pass in various curl options, including user agent string, whether to get verbose curl output or not, setting a timeout for requests, and more. See curl_options for all the options you can use.

Details

A progress helper will be coming soon.

Examples

# set curl options on client initialization
(res <- HttpClient$new(
  url = "https://httpbin.org",
  opts = list(
    verbose = TRUE,
    useragent = "hello world"
  )
))

#> <crul connection> 
#>   url: https://httpbin.org
#>   curl options: 
#>     verbose: TRUE
#>     useragent: hello world
#>   proxies: 
#>   headers: 

res$opts

#> $verbose
#> [1] TRUE
#> 
#> $useragent
#> [1] "hello world"
#> 

res$get('get')

#> <crul response> 
#>   url: https://httpbin.org/get
#>   request_headers: 
#>     useragent: hello world
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     server: nginx
#>     date: Wed, 22 Feb 2017 18:41:20 GMT
#>     content-type: application/json
#>     content-length: 229
#>     connection: keep-alive
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>   status: 200

# or set curl options when performing HTTP operation
(res <- HttpClient$new(url = "https://httpbin.org"))

#> <crul connection> 
#>   url: https://httpbin.org
#>   curl options: 
#>   proxies: 
#>   headers: 

res$get('get', verbose = TRUE)

#> <crul response> 
#>   url: https://httpbin.org/get
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.0.9100
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     server: nginx
#>     date: Wed, 22 Feb 2017 18:41:20 GMT
#>     content-type: application/json
#>     content-length: 259
#>     connection: keep-alive
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>   status: 200

## Not run: res$get(&#39;get&#39;, stuff = &quot;things&quot;)
## Not run: res$get(&#39;get&#39;, httpget = TRUE)

# set a timeout
(res <- HttpClient$new(
  url = "https://httpbin.org",
  opts = list(timeout_ms = 1)
))

#> <crul connection> 
#>   url: https://httpbin.org
#>   curl options: 
#>     timeout_ms: 1
#>   proxies: 
#>   headers: 

## Not run: res$get(&#39;get&#39;)