verb-HEAD {crul}	R Documentation

HTTP verb info: HEAD

Description

The HEAD method asks for a response identical to that of a GET request, but without the response body.

The HEAD method

The HEAD method is identical to GET except that the server MUST NOT send a message body in the response (i.e., the response terminates at the end of the header section). The server SHOULD send the same header fields in response to a HEAD request as it would have sent if the request had been a GET, except that the payload header fields MAY be omitted. This method can be used for obtaining metadata about the selected representation without transferring the representation data and is often used for testing hypertext links for validity, accessibility, and recent modification.

See https://tools.ietf.org/html/rfc7231#section-4.3.2 for further details.

References

https://tools.ietf.org/html/rfc7231#section-4.3.2

See Also

crul-package

Other verbs: verb-DELETE, verb-GET, verb-PATCH, verb-POST, verb-PUT

Examples

## Not run: 
x <- HttpClient$new(url = "https://httpbin.org")
x$head()

## End(Not run)

[Package crul version 0.7.0 Index]