verb-PATCH {crul}	R Documentation

HTTP verb info: PATCH

Description

The PATCH method is used to apply partial modifications to a resource.

The PATCH method

The PATCH method requests that a set of changes described in the request entity be applied to the resource identified by the Request- URI. The set of changes is represented in a format called a "patch document" identified by a media type. If the Request-URI does not point to an existing resource, the server MAY create a new resource, depending on the patch document type (whether it can logically modify a null resource) and permissions, etc.

See https://tools.ietf.org/html/rfc5789#section-2 for further details.

References

https://tools.ietf.org/html/rfc5789

See Also

crul-package

Other verbs: verb-DELETE, verb-GET, verb-HEAD, verb-POST, verb-PUT

Examples

## Not run: 
x <- HttpClient$new(url = "https://httpbin.org")
x$patch(path = 'patch', body = list(hello = "mars"))

## End(Not run)

[Package crul version 0.7.0 Index]