HTTP client

HTTP client

Arguments

url

(character) A url. One of url or handle required.

opts

(list) curl options, a named list. See curl_options for available curl options

proxies

an object of class proxy, as returned from the proxy function. Supports one proxy for now

headers

(list) a named list of headers

handle

A handle, see handle

Details

Methods

get(path, query, disk, stream, ...)

Make a GET request

post(path, query, body, disk, stream, ...)

Make a POST request

put(path, query, body, disk, stream, ...)

Make a PUT request

patch(path, query, body, disk, stream, ...)

Make a PATCH request

delete(path, query, body, disk, stream, ...)

Make a DELETE request

head(path, disk, stream, ...)

Make a HEAD request

Possible parameters (not all are allowed in each HTTP verb):

path - URL path, appended to the base URL

query - query terms, as a list

body - body as an R list

encode - one of form, multipart, json, or raw

disk - a path to write to. if NULL (default), memory used. See curl::curl_fetch_disk() for help.

stream - an R function to determine how to stream data. if NULL (default), memory used. See curl::curl_fetch_stream() for help

... curl options, only those in the acceptable set from curl::curl_options() except the following: httpget, httppost, post, postfields, postfieldsize, and customrequest

See also

post-requests, delete-requests, http-headers, writing-options

Examples

(x <- HttpClient$new(url = "https://httpbin.org"))

#> <crul connection> 
#>   url: https://httpbin.org
#>   curl options: 
#>   proxies: 
#>   headers: 

x$url

#> [1] "https://httpbin.org"

(res_get1 <- x$get('get'))

#> <crul response> 
#>   url: https://httpbin.org/get
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: gunicorn/19.7.1
#>     date: Fri, 24 Mar 2017 23:05:30 GMT
#>     content-type: application/json
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     content-length: 287
#>     via: 1.1 vegur
#>   status: 200

res_get1$content

#>   [1] 7b 0a 20 20 22 61 72 67 73 22 3a 20 7b 7d 2c 20 0a 20 20 22 68 65 61 64 65
#>  [26] 72 73 22 3a 20 7b 0a 20 20 20 20 22 41 63 63 65 70 74 22 3a 20 22 2a 2f 2a
#>  [51] 22 2c 20 0a 20 20 20 20 22 41 63 63 65 70 74 2d 45 6e 63 6f 64 69 6e 67 22
#>  [76] 3a 20 22 67 7a 69 70 2c 20 64 65 66 6c 61 74 65 22 2c 20 0a 20 20 20 20 22
#> [101] 43 6f 6e 6e 65 63 74 69 6f 6e 22 3a 20 22 63 6c 6f 73 65 22 2c 20 0a 20 20
#> [126] 20 20 22 48 6f 73 74 22 3a 20 22 68 74 74 70 62 69 6e 2e 6f 72 67 22 2c 20
#> [151] 0a 20 20 20 20 22 55 73 65 72 2d 41 67 65 6e 74 22 3a 20 22 6c 69 62 63 75
#> [176] 72 6c 2f 37 2e 35 31 2e 30 20 72 2d 63 75 72 6c 2f 32 2e 33 20 63 72 75 6c
#> [201] 2f 30 2e 33 2e 32 2e 39 34 31 34 22 0a 20 20 7d 2c 20 0a 20 20 22 6f 72 69
#> [226] 67 69 6e 22 3a 20 22 31 35 37 2e 31 33 30 2e 31 37 39 2e 38 36 22 2c 20 0a
#> [251] 20 20 22 75 72 6c 22 3a 20 22 68 74 74 70 73 3a 2f 2f 68 74 74 70 62 69 6e
#> [276] 2e 6f 72 67 2f 67 65 74 22 0a 7d 0a

res_get1$response_headers

#> $status
#> [1] "HTTP/1.1 200 OK"
#> 
#> $connection
#> [1] "keep-alive"
#> 
#> $server
#> [1] "gunicorn/19.7.1"
#> 
#> $date
#> [1] "Fri, 24 Mar 2017 23:05:30 GMT"
#> 
#> $`content-type`
#> [1] "application/json"
#> 
#> $`access-control-allow-origin`
#> [1] "*"
#> 
#> $`access-control-allow-credentials`
#> [1] "true"
#> 
#> $`content-length`
#> [1] "287"
#> 
#> $via
#> [1] "1.1 vegur"
#> 

res_get1$parse()

#> No encoding supplied: defaulting to UTF-8.

#> [1] "{\n  \"args\": {}, \n  \"headers\": {\n    \"Accept\": \"*/*\", \n    \"Accept-Encoding\": \"gzip, deflate\", \n    \"Connection\": \"close\", \n    \"Host\": \"httpbin.org\", \n    \"User-Agent\": \"libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414\"\n  }, \n  \"origin\": \"157.130.179.86\", \n  \"url\": \"https://httpbin.org/get\"\n}\n"

(res_get2 <- x$get('get', query = list(hello = "world")))

#> <crul response> 
#>   url: https://httpbin.org/get?hello=world
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: gunicorn/19.7.1
#>     date: Fri, 24 Mar 2017 23:05:30 GMT
#>     content-type: application/json
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     content-length: 323
#>     via: 1.1 vegur
#>   params: 
#>     hello: world
#>   status: 200

res_get2$parse()

#> No encoding supplied: defaulting to UTF-8.

#> [1] "{\n  \"args\": {\n    \"hello\": \"world\"\n  }, \n  \"headers\": {\n    \"Accept\": \"*/*\", \n    \"Accept-Encoding\": \"gzip, deflate\", \n    \"Connection\": \"close\", \n    \"Host\": \"httpbin.org\", \n    \"User-Agent\": \"libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414\"\n  }, \n  \"origin\": \"157.130.179.86\", \n  \"url\": \"https://httpbin.org/get?hello=world\"\n}\n"

library("jsonlite")
jsonlite::fromJSON(res_get2$parse())

#> No encoding supplied: defaulting to UTF-8.

#> $args
#> $args$hello
#> [1] "world"
#> 
#> 
#> $headers
#> $headers$Accept
#> [1] "*/*"
#> 
#> $headers$`Accept-Encoding`
#> [1] "gzip, deflate"
#> 
#> $headers$Connection
#> [1] "close"
#> 
#> $headers$Host
#> [1] "httpbin.org"
#> 
#> $headers$`User-Agent`
#> [1] "libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414"
#> 
#> 
#> $origin
#> [1] "157.130.179.86"
#> 
#> $url
#> [1] "https://httpbin.org/get?hello=world"
#> 

# post request
(res_post <- x$post('post', body = list(hello = "world")))

#> <crul response> 
#>   url: https://httpbin.org/post
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: gunicorn/19.7.1
#>     date: Fri, 24 Mar 2017 23:05:30 GMT
#>     content-type: application/json
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     content-length: 531
#>     via: 1.1 vegur
#>   status: 200

## empty body request
x$post('post')

#> <crul response> 
#>   url: https://httpbin.org/post
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: gunicorn/19.7.1
#>     date: Fri, 24 Mar 2017 23:05:30 GMT
#>     content-type: application/json
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     content-length: 437
#>     via: 1.1 vegur
#>   status: 200

# put request
(res_put <- x$put('put'))

#> <crul response> 
#>   url: https://httpbin.org/put
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: gunicorn/19.7.1
#>     date: Fri, 24 Mar 2017 23:05:30 GMT
#>     content-type: application/json
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     content-length: 436
#>     via: 1.1 vegur
#>   status: 200

# delete request
(res_delete <- x$delete('delete'))

#> <crul response> 
#>   url: https://httpbin.org/delete
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: gunicorn/19.7.1
#>     date: Fri, 24 Mar 2017 23:05:30 GMT
#>     content-type: application/json
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     content-length: 439
#>     via: 1.1 vegur
#>   status: 200

# patch request
(res_patch <- x$patch('patch'))

#> <crul response> 
#>   url: https://httpbin.org/patch
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: gunicorn/19.7.1
#>     date: Fri, 24 Mar 2017 23:05:30 GMT
#>     content-type: application/json
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     content-length: 438
#>     via: 1.1 vegur
#>   status: 200

# head request
(res_head <- x$head())

#> <crul response> 
#>   url: https://httpbin.org/
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.2.9414
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: gunicorn/19.7.1
#>     date: Fri, 24 Mar 2017 23:05:30 GMT
#>     content-type: text/html; charset=utf-8
#>     content-length: 12172
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     via: 1.1 vegur
#>   status: 200

# query params are URL encoded for you, so DO NOT do it yourself
## if you url encode yourself, it gets double encoded, and that's bad
(x <- HttpClient$new(url = "https://httpbin.org"))

#> <crul connection> 
#>   url: https://httpbin.org
#>   curl options: 
#>   proxies: 
#>   headers: 

res <- x$get("get", query = list(a = 'hello world'), verbose = TRUE)