HTTP request object

HTTP request object

Arguments

url

(character) A url. One of url or handle required.

opts

(list) curl options, a named list. See curl_options for available curl options

proxies

an object of class proxy, as returned from the proxy function. Supports one proxy for now

headers

(list) a named list of headers

handle

A handle, see handle

Details

This R6 class doesn't do actual HTTP requests as does HttpClient - but is for building requests to use for async HTTP requests in AsyncVaried

Methods

get(path, query, disk, stream, ...)

Define a GET request

post(path, query, body, disk, stream, ...)

Define a POST request

put(path, query, body, disk, stream, ...)

Define a PUT request

patch(path, query, body, disk, stream, ...)

Define a PATCH request

delete(path, query, body, disk, stream, ...)

Define a DELETE request

head(path, disk, stream, ...)

Define a HEAD request

See HttpClient for information on parameters.

See also

post-requests, http-headers, writing-options

Examples

(x <- HttpRequest$new(url = "https://httpbin.org/get")$get())

#> <crul http request> 
#>   url: https://httpbin.org/get
#>   curl options: 
#>   proxies: 
#>   headers: 

x$url

#> [1] "https://httpbin.org/get"

x$payload

#> $url
#> $url$url
#> [1] "https://httpbin.org/get"
#> 
#> $url$handle
#> <curl handle> (https://httpbin.org/get)
#> 
#> 
#> $method
#> [1] "get"
#> 
#> $options
#> $options$httpget
#> [1] TRUE
#> 
#> $options$useragent
#> [1] "libcurl/7.51.0 r-curl/2.3 crul/0.3.0.9100"
#> 
#> 
#> $headers
#> list()
#> 

(x <- HttpRequest$new(url = "https://httpbin.org/post"))

#> <crul http request> 
#>   url: https://httpbin.org/post
#>   curl options: 
#>   proxies: 
#>   headers: 

x$post(body = list(foo = "bar"))

#> <crul http request> 
#>   url: https://httpbin.org/post
#>   curl options: 
#>   proxies: 
#>   headers: 

HttpRequest$new(
  url = "https://httpbin.org/get",
  headers = list(
    `Content-Type` = "application/json"
  )
)

#> <crul http request> 
#>   url: https://httpbin.org/get
#>   curl options: 
#>   proxies: 
#>   headers: 
#>     Content-Type: application/json