Base response object

Base response object

Arguments

url
(character) A url, required

opts
(list) curl options

handle
A handle

method
(character) HTTP method

status_code
(integer) status code

request_headers
(list) request headers, named list

response_headers
(list) response headers, named list

modified
(character) modified date

times
(vector) named vector

content
(raw) raw binary content response

request
request object, with all details

Details

Methods

parse()

Parse the raw response content to text

success()

Was status code less than or equal to 201. returns boolean

status_http()

Get HTTP status code, message, and explanation

raise_for_status()

Check HTTP status and stop with appropriate HTTP error code and message if >= 300. - If you have fauxpas installed we use that, otherwise use httpcode

Examples

x <- HttpResponse$new(method = "get", url = "https://httpbin.org")
x$url

#> [1] "https://httpbin.org"

x$method

#> [1] "get"

x <- HttpClient$new(url = 'https://httpbin.org')
(res <- x$get('get'))

#> <crul response> 
#>   url: https://httpbin.org/get
#>   request_headers: 
#>     User-Agent: libcurl/7.51.0 r-curl/2.6 crul/0.3.5.9110
#>     Accept-Encoding: gzip, deflate
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     connection: keep-alive
#>     server: meinheld/0.6.1
#>     date: Mon, 15 May 2017 18:12:01 GMT
#>     content-type: application/json
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     x-powered-by: Flask
#>     x-processed-time: 0.00114893913269
#>     content-length: 287
#>     via: 1.1 vegur
#>   status: 200

res$request_headers

#> $`User-Agent`
#> [1] "libcurl/7.51.0 r-curl/2.6 crul/0.3.5.9110"
#> 
#> $`Accept-Encoding`
#> [1] "gzip, deflate"
#> 

res$response_headers

#> $status
#> [1] "HTTP/1.1 200 OK"
#> 
#> $connection
#> [1] "keep-alive"
#> 
#> $server
#> [1] "meinheld/0.6.1"
#> 
#> $date
#> [1] "Mon, 15 May 2017 18:12:01 GMT"
#> 
#> $`content-type`
#> [1] "application/json"
#> 
#> $`access-control-allow-origin`
#> [1] "*"
#> 
#> $`access-control-allow-credentials`
#> [1] "true"
#> 
#> $`x-powered-by`
#> [1] "Flask"
#> 
#> $`x-processed-time`
#> [1] "0.00114893913269"
#> 
#> $`content-length`
#> [1] "287"
#> 
#> $via
#> [1] "1.1 vegur"
#> 

res$parse()

#> No encoding supplied: defaulting to UTF-8.

#> [1] "{\n  \"args\": {}, \n  \"headers\": {\n    \"Accept\": \"*/*\", \n    \"Accept-Encoding\": \"gzip, deflate\", \n    \"Connection\": \"close\", \n    \"Host\": \"httpbin.org\", \n    \"User-Agent\": \"libcurl/7.51.0 r-curl/2.6 crul/0.3.5.9110\"\n  }, \n  \"origin\": \"157.130.179.86\", \n  \"url\": \"https://httpbin.org/get\"\n}\n"

res$status_code

#> [1] 200

res$status_http()

#> <Status code: 200>
#>   Message: OK
#>   Explanation: Request fulfilled, document follows

res$status_http()$status_code

#> [1] "200"

res$status_http()$message

#> [1] "OK"

res$status_http()$explanation

#> [1] "Request fulfilled, document follows"

res$success()

#> [1] TRUE

x <- HttpClient$new(url = 'https://httpbin.org/status/404')
(res <- x$get())

#> <crul response> 
#>   url: https://httpbin.org/status/404
#>   request_headers: 
#>     User-Agent: libcurl/7.51.0 r-curl/2.6 crul/0.3.5.9110
#>     Accept-Encoding: gzip, deflate
#>   response_headers: 
#>     status: HTTP/1.1 404 NOT FOUND
#>     connection: keep-alive
#>     server: meinheld/0.6.1
#>     date: Mon, 15 May 2017 18:12:01 GMT
#>     content-type: text/html; charset=utf-8
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     x-powered-by: Flask
#>     x-processed-time: 0.000746965408325
#>     content-length: 0
#>     via: 1.1 vegur
#>   status: 404

 ## Not run: res$raise_for_status()

x <- HttpClient$new(url = 'https://httpbin.org/status/414')
(res <- x$get())

#> <crul response> 
#>   url: https://httpbin.org/status/414
#>   request_headers: 
#>     User-Agent: libcurl/7.51.0 r-curl/2.6 crul/0.3.5.9110
#>     Accept-Encoding: gzip, deflate
#>   response_headers: 
#>     status: HTTP/1.1 414 REQUEST URI TOO LONG
#>     connection: keep-alive
#>     server: meinheld/0.6.1
#>     date: Mon, 15 May 2017 18:12:01 GMT
#>     content-type: text/html; charset=utf-8
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>     x-powered-by: Flask
#>     x-processed-time: 0.000998020172119
#>     content-length: 0
#>     via: 1.1 vegur
#>   status: 414

 ## Not run: res$raise_for_status()