Base response object

Base response object

Arguments

url

(character) A url

opts

(list) curl options

handle

A handle

Details

Methods

parse()

Parse the raw response content to text

success()

Was status code less than or equal to 201. returns boolean

status_http()

Get HTTP status code, message, and explanation

raise_for_status()

Check HTTP status and stop with appropriate HTTP error code and message if >= 300. - If you have fauxpas installed we use that, otherwise use httpcode

Examples

x <- HttpResponse$new(method = "get", url = "https://httpbin.org")
x$url

#> [1] "https://httpbin.org"

x$method

#> [1] "get"

x <- HttpClient$new(url = 'https://httpbin.org')
(res <- x$get('get'))

#> <crul response> 
#>   url: https://httpbin.org/get
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.0.9100
#>   response_headers: 
#>     status: HTTP/1.1 200 OK
#>     server: nginx
#>     date: Wed, 22 Feb 2017 18:41:23 GMT
#>     content-type: application/json
#>     content-length: 259
#>     connection: keep-alive
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>   status: 200

res$request_headers

#> $useragent
#> [1] "libcurl/7.51.0 r-curl/2.3 crul/0.3.0.9100"
#> 

res$response_headers

#> $status
#> [1] "HTTP/1.1 200 OK"
#> 
#> $server
#> [1] "nginx"
#> 
#> $date
#> [1] "Wed, 22 Feb 2017 18:41:23 GMT"
#> 
#> $`content-type`
#> [1] "application/json"
#> 
#> $`content-length`
#> [1] "259"
#> 
#> $connection
#> [1] "keep-alive"
#> 
#> $`access-control-allow-origin`
#> [1] "*"
#> 
#> $`access-control-allow-credentials`
#> [1] "true"
#> 

res$parse()

#> No encoding supplied: defaulting to UTF-8.

#> [1] "{\n  \"args\": {}, \n  \"headers\": {\n    \"Accept\": \"*/*\", \n    \"Accept-Encoding\": \"gzip, deflate\", \n    \"Host\": \"httpbin.org\", \n    \"User-Agent\": \"libcurl/7.51.0 r-curl/2.3 crul/0.3.0.9100\"\n  }, \n  \"origin\": \"157.130.179.86\", \n  \"url\": \"https://httpbin.org/get\"\n}\n"

res$status_code

#> [1] 200

res$status_http()

#> <Status code: 200>
#>   Message: OK
#>   Explanation: Request fulfilled, document follows

res$status_http()$status_code

#> [1] "200"

res$status_http()$message

#> [1] "OK"

res$status_http()$explanation

#> [1] "Request fulfilled, document follows"

res$success()

#> [1] TRUE

x <- HttpClient$new(url = 'https://httpbin.org/status/404')
(res <- x$get())

#> <crul response> 
#>   url: https://httpbin.org/status/404
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.0.9100
#>   response_headers: 
#>     status: HTTP/1.1 404 NOT FOUND
#>     server: nginx
#>     date: Wed, 22 Feb 2017 18:41:24 GMT
#>     content-type: text/html; charset=utf-8
#>     content-length: 0
#>     connection: keep-alive
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>   status: 404

 ## Not run: res$raise_for_status()

x <- HttpClient$new(url = 'https://httpbin.org/status/414')
(res <- x$get())

#> <crul response> 
#>   url: https://httpbin.org/status/414
#>   request_headers: 
#>     useragent: libcurl/7.51.0 r-curl/2.3 crul/0.3.0.9100
#>   response_headers: 
#>     status: HTTP/1.1 414 REQUEST URI TOO LONG
#>     server: nginx
#>     date: Wed, 22 Feb 2017 18:41:24 GMT
#>     content-type: text/html; charset=utf-8
#>     content-length: 0
#>     connection: keep-alive
#>     access-control-allow-origin: *
#>     access-control-allow-credentials: true
#>   status: 414

 ## Not run: res$raise_for_status()