crul-package {crul}	R Documentation

HTTP R client

Description

HTTP R client

Package API

HttpClient() - create a connection client, set all your http options, make http requests

HttpResponse() - mostly for internal use, handles http responses

Paginator() - auto-paginate through requests

Async() - asynchronous requests

AsyncVaried() - varied asynchronous requests

HttpRequest() - generate an HTTP request, mostly for use in building requests to be used in Async or AsyncVaried

mock() - Turn on/off mocking, via webmockr

auth() - Simple authentication helper

proxy() - Proxy helper

HTTP conditions

We use fauxpas if you have it installed for handling HTTP conditions but if it's not installed we use httpcode

Mocking

Mocking HTTP requests is supported via the webmockr package. See mock for guidance.

Author(s)

Scott Chamberlain myrmecocystus@gmail.com

[Package crul version 0.5.0 Index]