verb-DELETE {crul}	R Documentation

HTTP verb info: DELETE

Description

The DELETE method deletes the specified resource.

The DELETE method

The DELETE method requests that the origin server remove the association between the target resource and its current functionality. In effect, this method is similar to the rm command in UNIX: it expresses a deletion operation on the URI mapping of the origin server rather than an expectation that the previously associated information be deleted.

See https://tools.ietf.org/html/rfc7231#section-4.3.5 for further details.

References

https://tools.ietf.org/html/rfc7231#section-4.3.5

See Also

crul-package

Other verbs: verb-GET, verb-HEAD, verb-PATCH, verb-POST, verb-PUT

Examples

## Not run: 
x <- HttpClient$new(url = "https://httpbin.org")
x$delete(path = 'delete')

## End(Not run)

[Package crul version 0.7.0 Index]