verb-POST {crul}	R Documentation

HTTP verb info: POST

Description

The POST method is used to submit an entity to the specified resource, often causing a change in state or side effects on the server.

The POST method

If one or more resources has been created on the origin server as a result of successfully processing a POST request, the origin server SHOULD send a 201 (Created) response containing a Location header field that provides an identifier for the primary resource created (Section 7.1.2 https://tools.ietf.org/html/rfc7231#section-7.1.2) and a representation that describes the status of the request while referring to the new resource(s).

See https://tools.ietf.org/html/rfc7231#section-4.3.3 for further details.

References

https://tools.ietf.org/html/rfc7231#section-4.3.3

See Also

crul-package

Other verbs: verb-DELETE, verb-GET, verb-HEAD, verb-PATCH, verb-PUT

Examples

## Not run: 
x <- HttpClient$new(url = "https://httpbin.org")
x$post(path = 'post', body = list(hello = "world"))

## End(Not run)

[Package crul version 0.8.0 Index]